build: make nacl pass
Add nacl.bash, the NaCl version of all.bash.
It's a separate script because it builds a variant of package syscall
with a large zip file embedded in it, containing all the input files
needed for tests.

Disable various tests new since the last round, mostly the ones using os/exec.

Fixes #7945.

LGTM=dave
R=golang-codereviews, remyoudompheng, dave, bradfitz
CC=golang-codereviews
https://golang.org/cl/100590044

+++ test/gc2.go
@@ -1,5 +1,7 @@
 // run
 
+// +build !nacl
+
 // Copyright 2011 The Go Authors.  All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
@@ -36,7 +38,7 @@ func main() {
 	}
 
 	runtime.ReadMemStats(memstats)
-	obj := memstats.HeapObjects - st.HeapObjects
+	obj := int64(memstats.HeapObjects - st.HeapObjects)
 	if obj > N/5 {
 		fmt.Println("too many objects left:", obj)
 		os.Exit(1)
